Fix: position fixed 부모기준
1. 부모 너비 찾아서 그만큼 width값 설정
=> 훅으로 분리할 것
