+++ src/components/Common/TextEdit.tsx
@@ -1,6 +1,6 @@
 import { useForm } from "react-hook-form";
 import { SendHorizontal } from "lucide-react";
-import { FormEvent } from "react";
+import { FormEvent, useEffect, useState } from "react";
 
 import useUploadPost from "@/hooks/useUploadPost.tsx";
 import useUploadComment from "@/hooks/useUploadComment.tsx";
@@ -8,10 +8,12 @@ import useUploadComment from "@/hooks/useUploadComment.tsx";
 import { Textarea } from "@/components/ui/textarea.tsx";
 
 import { cn } from "@/lib/utils";
+import MentionInput from "@/components/Common/Mention/MentionInput.tsx";
 
 export type ContentType = "post" | "comment";
 export type SubmitType = "create" | "patch";
 interface Props {
+  isMention: boolean;
   contentType: ContentType;
   submitType: SubmitType;
   userName: string | undefined;
@@ -19,7 +21,7 @@ interface Props {
   channelId: string;
 }
 
-const TextEdit = ({ contentType, submitType, userName, postId, channelId }: Props) => {
+const TextEdit = ({ isMention, contentType, submitType, userName, postId, channelId }: Props) => {
   const { register, handleSubmit, watch, setValue } = useForm({
     defaultValues: { anonymous: true, content: "" },
   });
@@ -42,37 +44,46 @@ const TextEdit = ({ contentType, submitType, userName, postId, channelId }: Prop
     }
   };
 
+  // todo 24/1/2 부모 너비 가져오기
+  const [parentWidth, setParentWidth] = useState(0);
+  useEffect(() => {
+    setParentWidth(document.getElementById("parent")?.offsetWidth);
+  }, []);
+
   return (
-    <form className="relative">
-      <Textarea
-        placeholder={`${contentType}을 작성해주세요.`}
-        className="resize-none"
-        {...register("content")}
-      />
-      <div className="absolute bottom-2 right-2 flex items-center gap-2">
-        <label
-          className="flex cursor-pointer items-center gap-2 rounded-xl border p-3"
-          htmlFor="anonymous"
-        >
-          <input
-            type="checkbox"
-            id="anonymous"
-            {...register("anonymous")}
-            onClick={handleCheckClick}
-          />
-          <p className="text-gray-500">익명</p>
-        </label>
-        <button
-          onClick={handleSubmit(handleUpload)}
-          className={cn(
-            "flex h-12 w-12 cursor-pointer items-center justify-center rounded-xl text-black",
-            watch("content") ? "bg-green-600" : "",
-          )}
-        >
-          <SendHorizontal className="h-10 w-10 fill-amber-50 stroke-2" />
-        </button>
-      </div>
-    </form>
+    <div className="fixed bottom-0" style={{ width: parentWidth }}>
+      {isMention && <MentionInput />}
+      <form className="relative">
+        <Textarea
+          placeholder={`${contentType}을 작성해주세요.`}
+          className="resize-none"
+          {...register("content")}
+        />
+        <div className="absolute bottom-2 right-2 flex items-center gap-2">
+          <label
+            className="flex cursor-pointer items-center gap-2 rounded-xl border p-3"
+            htmlFor="anonymous"
+          >
+            <input
+              type="checkbox"
+              id="anonymous"
+              {...register("anonymous")}
+              onClick={handleCheckClick}
+            />
+            <p className="text-gray-500">익명</p>
+          </label>
+          <button
+            onClick={handleSubmit(handleUpload)}
+            className={cn(
+              "flex h-12 w-12 cursor-pointer items-center justify-center rounded-xl text-black",
+              watch("content") ? "bg-green-600" : "",
+            )}
+          >
+            <SendHorizontal className="h-10 w-10 fill-amber-50 stroke-2" />
+          </button>
+        </div>
+      </form>
+    </div>
   );
 };
