--- src/components/Demo/ExMentionInput.tsx
@@ -1,11 +1,0 @@
-import MentionInput from "@/components/Common/Mention/MentionInput.tsx";
-
-const ExMentionInput = () => {
-  return (
-    <div className="p-20">
-      <MentionInput />
-    </div>
-  );
-};
-
-export default ExMentionInput;
